ai/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agri-assistant-nlu"
version = "0.1.0"
description = "Minimal offline joint intent+slot NLU for an agri assistant"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Agri NLU" }]
dependencies = [
  "pyyaml>=6.0",
  "click>=8.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
train = [
  "torch>=2.2",
  "transformers>=4.41",
  "numpy>=1.26",
  "tqdm>=4.66",
  "regex>=2024.5.15",
  "scikit-learn>=1.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
  "ignore::UserWarning",
]

[tool.setuptools]
packages = ["nlu"]

